requirements.txt
python-dotenv==1.0.1
lxml==5.3.1
python-docx==1.1.2
groq